+++ gateway/src/redis_io/mod.rs
@@ -1,10 +1,12 @@
 pub mod entry_id_repo;
+mod into_shared;
 mod key_provider;
 mod namespace;
 pub mod stream;
 pub mod xadd;
 pub mod xread;
 
+pub use into_shared::IntoShared;
 pub use key_provider::*;
 pub use namespace::*;
 pub use xadd::start;
